Solving this assignment will give us an idea about how problems are systematically solved using Exploratory Data Analysis(EDA) and data visualization.
Business Understanding :
We all have used Uber or any other cab service for travel from the airport. Most of the time problem of cancellation by the driver or non-availability of cars faced by the customers. These issues also impact the business of Uber as Uber loses out on its revenue. As an analyst, we decide to address the problem Uber is facing - driver cancellation and non-availability of cars leading to loss of potential revenue.

Business Objectives :
The aim of analysis is to identify the root cause of the problem (i.e. cancellation and non-availability of cars) and recommend ways to improve the situation. As a result of your analysis, you should be able to present to the client the root cause(s) and possible hypotheses of the problem(s) and recommend ways to improve them.

Data Understanding :

There are six attributes associated with each request made by a customer:
           1. Request id: A unique identifier of the request
           2. Time of request: The date and time at which the customer made the trip request
           3. Drop-off time: The drop-off date and time, in case the trip was completed
           4. Pick-up point: The point from which the request was made
           5. Driver id: The unique identification number of the driver
           6. Status of the request: The final status of the trip, that can be either completed, cancelled by the driver or no cars available

Note: For this assignment, only the trips to and from the airport are being considered.